Fix codeblock rendering
The codeblocks returned by lintian-explain-tags are not exactly
markdown, but instead they are indented with simply 1 space and they
contain HTML entities that should not be escaped. Thus we need to extend
the parser to allow code block with a different indentation. Then we add
a node renderer that do not escape its content chen rendering it to
HTML.

We also exclude the headings block parser, as we never want any headings
in the parsed text.