Disable Linkify extention for markdown.StyleFull
As it currently renders incorrectly some links that were probably not intended
to be linkifyed. It is better to only convert to HTML links those that were
explicitely marked as so by surrounding them with <>.

By the way, rename markdown testdata to indicate explicitely what they are supposed
to test, and also add a comment to the source tag from which this output originaly
come.